run axe a11y tests sequentially
