+++ app/src/renderer/views/Inbox.test.tsx
@@ -1,19 +1,7 @@
 import { screen } from "@testing-library/react";
 import { expect } from "vitest";
 import InboxView from "./Inbox";
-import {
-  renderWithProviders,
-  renderAndCheckA11y,
-} from "../test-component-setup";
-
-describe("InboxView accessibility", () => {
-  it("has no axe violations on initial render", async () => {
-    // TODO(a11y): ARIA input field is missing an accessible name.
-    await renderAndCheckA11y(<InboxView />, undefined, [
-      "aria-input-field-name",
-    ]);
-  });
-});
+import { renderWithProviders } from "../test-component-setup";
 
 describe("InboxView Component", () => {
   it('says the string "Select a source"', () => {
