+++ app/src/renderer/views/Inbox/Sidebar/SourceList/Source.test.tsx
@@ -3,7 +3,6 @@ import { expect, describe, it, vi } from "vitest";
 import userEvent from "@testing-library/user-event";
 import {
   renderWithProviders,
-  renderAndCheckA11y,
   testMemoization,
 } from "../../../../test-component-setup";
 import Source from "./Source";
@@ -49,24 +48,6 @@ const defaultProps = {
   onSelect: mockOnSelect,
   onToggleStar: mockOnToggleStar,
 };
-
-describe("Source accessibility", () => {
-  it("has no axe violations on a basic unread source", async () => {
-    // TODO(a11y): source row has unnamed buttons, unlabeled form controls,
-    // nested interactive controls, and an ARIA role outside its required parent.
-    await renderAndCheckA11y(
-      <Source
-        source={createMockSource()}
-        isSelected={false}
-        isActive={false}
-        onSelect={vi.fn()}
-        onToggleStar={vi.fn()}
-      />,
-      undefined,
-      ["aria-required-parent", "button-name", "label", "nested-interactive"],
-    );
-  });
-});
 
 describe("Source Component", () => {
   beforeEach(() => {
